+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/analytics/EventAttributeName.kt
@@ -84,6 +84,11 @@ enum class EventAttributeName {
     maxVideoTileCount,
 
     /**
+     * Duration of the meeting starting process
+     */
+    meetingStartDurationMs,
+
+    /**
      * Duration of the meeting
      */
     meetingDurationMs,